move for CI mem issues (see 117eb0b)
