+++ src/services/goalServices/getGoalsForReport.test.js
@@ -2,7 +2,7 @@ import faker from '@faker-js/faker';
 import { REPORT_STATUSES } from '@ttahub/common';
 import {
   getGoalsForReport,
-} from '../goals';
+} from './goals';
 import {
   Goal,
   Objective,
